made stripe function modularized, using req.body.phone for twilio instead of hard coding the phone number, gitignored twilio phone number
